refactor(main): update routes and middleware to use shared utilities
