+++ src/main/fastify/routes/answers/delete-answer.route.ts
@@ -4,7 +4,7 @@ import {
   deleteAnswerResponsesSchema
 } from '@/infra/validation/zod/schemas/presentation/answers/delete-answer.schemas'
 import { makeDeleteAnswerController } from '@/main/factories/delete-answer'
-import { adaptRoute } from '@/util/adapt-route'
+import { adaptRoute } from '@/shared/util/http/adapt-route'
 
 export async function deleteAnswerRoute (app: FastifyInstance, tags: string[]) {
   app.delete('/:answerId', {
