+++ src/main/fastify/routes/comments/fetch-answer-comments.route.ts
@@ -5,7 +5,7 @@ import {
   fetchAnswerCommentsResponsesSchema
 } from '@/infra/validation/zod/schemas/presentation/comments/fetch-answer-comments.schemas'
 import { makeFetchAnswerCommentsController } from '@/main/factories/fetch-answer-comments'
-import { adaptRoute } from '@/util/adapt-route'
+import { adaptRoute } from '@/shared/util/http/adapt-route'
 
 export async function fetchAnswerCommentsRoute (app: FastifyInstance, tags: string[]) {
   app.get('/:answerId/comments', {
