+++ src/main/fastify/routes/questions/choose-question-best-answer.route.ts
@@ -1,7 +1,7 @@
 import type { FastifyInstance } from 'fastify'
 import { chooseQuestionBestAnswerParamsSchema, chooseQuestionBestAnswerResponsesSchema } from '@/infra/validation/zod/schemas/presentation/questions/choose-question-best-answer.schemas'
 import { makeChooseQuestionBestAnswerController } from '@/main/factories/choose-question-best-answer'
-import { adaptRoute } from '@/util/adapt-route'
+import { adaptRoute } from '@/shared/util/http/adapt-route'
 
 export async function chooseQuestionBestAnswerRoute (app: FastifyInstance, tags: string[]) {
   app.patch('/:answerId/choose', {
