+++ src/main/fastify/routes/questions/create-question.route.ts
@@ -4,7 +4,7 @@ import {
   createQuestionResponsesSchema
 } from '@/infra/validation/zod/schemas/presentation/questions/create-question.schemas'
 import { makeCreateQuestionController } from '@/main/factories/create-question'
-import { adaptRoute } from '@/util/adapt-route'
+import { adaptRoute } from '@/shared/util/http/adapt-route'
 
 export async function createQuestionRoute (app: FastifyInstance, tags: string[]) {
   app.post('/', {
