+++ src/main/fastify/routes/questions/delete-question.route.ts
@@ -4,7 +4,7 @@ import {
   deleteQuestionResponsesSchema
 } from '@/infra/validation/zod/schemas/presentation/questions/delete-question.schemas'
 import { makeDeleteQuestionController } from '@/main/factories/delete-question'
-import { adaptRoute } from '@/util/adapt-route'
+import { adaptRoute } from '@/shared/util/http/adapt-route'
 
 export async function deleteQuestionRoute (app: FastifyInstance, tags: string[]) {
   app.delete('/:questionId', {
