+++ src/main/fastify/routes/questions/fetch-questions.route.ts
@@ -2,8 +2,8 @@ import type { FastifyInstance } from 'fastify'
 import { paginationParamsSchema } from '@/infra/validation/zod/schemas/core/pagination-params.schema'
 import { fetchQuestionsResponsesSchemas } from '@/infra/validation/zod/schemas/presentation/questions/fetch-questions.schemas'
 import { makeFetchQuestionsController } from '@/main/factories/fetch-questions'
-import { adaptRoute } from '@/util/adapt-route'
 import { readOperationsRateLimit } from '../../plugins/rate-limit'
+import { adaptRoute } from '@/shared/util/http/adapt-route'
 
 export async function fetchQuestionsRoute (app: FastifyInstance, tags: string[]) {
   app.get('/', {
