+++ src/main/fastify/routes/questions/get-question-by-slug.route.ts
@@ -4,8 +4,8 @@ import {
   getQuestionBySlugResponsesSchema
 } from '@/infra/validation/zod/schemas/presentation/questions/get-question-by-slug.schemas'
 import { makeGetQuestionBySlugController } from '@/main/factories/get-question-by-slug'
-import { adaptRoute } from '@/util/adapt-route'
 import { readOperationsRateLimit } from '../../plugins/rate-limit'
+import { adaptRoute } from '@/shared/util/http/adapt-route'
 
 export async function getQuestionBySlugRoute (app: FastifyInstance, tags: string[]) {
   app.get('/:slug', {
