+++ src/main/fastify/routes/users/delete-account.route.ts
@@ -3,7 +3,7 @@ import {
   deleteAccountResponsesSchema
 } from '@/infra/validation/zod/schemas/presentation/users/delete-account.schemas'
 import { makeDeleteAccountController } from '@/main/factories/delete-account'
-import { adaptRoute } from '@/util/adapt-route'
+import { adaptRoute } from '@/shared/util/http/adapt-route'
 
 export async function deleteAccountRoute (app: FastifyInstance, tags: string[]) {
   app.delete('/', {
