+++ src/main/fastify/routes/users/fetch-users.route.ts
@@ -2,8 +2,8 @@ import type { FastifyInstance } from 'fastify'
 import { paginationParamsSchema } from '@/infra/validation/zod/schemas/core/pagination-params.schema'
 import { fetchUsersResponsesSchemas } from '@/infra/validation/zod/schemas/presentation/users/fetch-users.schemas'
 import { makeFetchUsersController } from '@/main/factories/fetch-users'
-import { adaptRoute } from '@/util/adapt-route'
 import { readOperationsRateLimit } from '../../plugins/rate-limit'
+import { adaptRoute } from '@/shared/util/http/adapt-route'
 
 export async function fetchUsersRoute (app: FastifyInstance, tags: string[]) {
   app.get('', {
