+++ src/main/fastify/routes/users/verify-email-validation.route.ts
@@ -4,8 +4,8 @@ import {
   verifyEmailValidationResponsesSchema
 } from '@/infra/validation/zod/schemas/presentation/users/verify-email-validation.schemas'
 import { makeVerifyEmailValidationController } from '@/main/factories/verify-email-validation-controller'
-import { adaptRoute } from '@/util/adapt-route'
 import { emailValidationRateLimit } from '../../plugins/rate-limit'
+import { adaptRoute } from '@/shared/util/http/adapt-route'
 
 export async function verifyEmailValidationRoute (app: FastifyInstance, tags: string[]) {
   app.post('/verify-email-validation', {
